+++ tests/testthat/test-theme_elements_gtsummary.R
@@ -124,3 +124,33 @@ test_that("tbl_summary-arg:missing_stat works", {
     "7 / 98"
   )
 })
+
+# add_overall.tbl_summary-arg:col_label ----------------------------------------
+test_that("tbl_summary-arg:missing_stat works", {
+  expect_equal(
+    with_gtsummary_theme(
+      x = list("add_overall.tbl_summary-arg:col_label" = "All Participants  \nN = {style_number(N)}"),
+      expr =
+        tbl_summary(trial, include = age, by = trt) |>
+        add_overall(last = TRUE) |>
+        as.data.frame() |>
+        names() |>
+        dplyr::last()
+    ),
+    "All Participants  \nN = 200"
+  )
+
+  expect_equal(
+    with_gtsummary_theme(
+      x = list("add_overall.tbl_summary-arg:col_label" = "All Participants  \nN = {style_number(N)}"),
+      expr =
+        survey::svydesign(~1, data = trial, weights = ~1) |>
+        tbl_svysummary(by = trt, include = age) |>
+        add_overall(last = TRUE) |>
+        as.data.frame() |>
+        names() |>
+        dplyr::last()
+    ),
+    "All Participants  \nN = 200"
+  )
+})